Reversing the damage

+++ KvadratniKoren/Program.cs
@@ -7,8 +7,8 @@ namespace KvadratniKorenjenec
 		static void Main(string[] args)
         {
 			Console.WriteLine("Type the square root you are trying to get");
-			float korenjenec = float.Parse(Console.ReadLine());
-            float? korenjenoStevilo = SquareRoot.KvadratniKoren.Koren((float)korenjenec);
+			long korenjenec = Int64.Parse(Console.ReadLine());
+            long? korenjenoStevilo = SquareRoot.KvadratniKoren.Koren(korenjenec);
 			if(korenjenoStevilo == null)
             {
 				Console.WriteLine("Korenjenec {0} nima celega korena", korenjenec);
